Fallback to text if no AWS Credentials are found

+++ chatgpt-ko.py
@@ -4,6 +4,7 @@ import boto3
 import pyaudio
 import os
 import sys
+import configparser
 from loguru import logger
 from ctypes import *
 from contextlib import contextmanager
@@ -37,7 +38,7 @@ logger.remove()
 logger.add("/tmp/ai.log", format="{message}")
 
 # Variables
-openai.api_key = "API_KEY"
+openai.api_key = "API-KEY"
 
 # Select the engine
 print("사용하고자 하는 엔진을 선택하십시오.:")
@@ -87,7 +88,9 @@ else:
         else:
             break
 
-# Add a variable to control flite usage
+# Add a variable to control speech usage
+config = configparser.ConfigParser()
+config.read(os.path.expanduser('~/.aws/credentials'))
 use_speech = input("Text-to-Speech 텍스트 음성합성을 사용하시겠습니까? (y/n) (기본값: n) : ") or "n"
 if use_speech == "y":
     speech_engine = input("어떤 음성 엔진을 사용하시겠습니까? (neural or default: standard): ") or "standard"
@@ -139,33 +142,34 @@ while True:
     # Print the text to the console
     print("\033[35mAI:  \033[0m" + text)
 
-    # Speak the text using flite if use_flite is "y"
+    # Speak the text using AWS Polly if use_speech is "y"
     if use_speech == "y":
-        # subprocess.run(["flite", "-voice", "slt", "--setf", "duration_stretch=1.15", "--setf", "int_f0_target_mean=160", "-pw", "-t", text], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-        #
-        polly = boto3.client('polly')
-        response = polly.synthesize_speech(
-            Engine=speech_engine,
-            LanguageCode='ko-KR',
-            VoiceId='Seoyeon',
-            OutputFormat='pcm',
-            Text = text
-        )
+        if config.has_section('default'):
+            polly = boto3.client('polly')
+            response = polly.synthesize_speech(
+                Engine=speech_engine,
+                LanguageCode='ko-KR',
+                VoiceId='Seoyeon',
+                OutputFormat='pcm',
+                Text = text
+            )
         
-        audio_stream = response['AudioStream'].read()
-
-        with noalsaerr():
-            p = pyaudio.PyAudio()
-            stream = p.open(format=p.get_format_from_width(width=2),
-                            channels=1,
-                            rate=16000,
-                            output=True)
-            stream.start_stream()
-            stream.write(audio_stream)
-            stream.stop_stream()
-            stream.close()
-            p.terminate()
-
+            # Audio Steam Settings
+            audio_stream = response['AudioStream'].read()
+            with noalsaerr():
+                p = pyaudio.PyAudio()
+                stream = p.open(format=p.get_format_from_width(width=2),
+                                channels=1,
+                                rate=16000,
+                                output=True)
+                stream.start_stream()
+                stream.write(audio_stream)
+                stream.stop_stream()
+                stream.close()
+                p.terminate()
+        else:
+            print("AWS credentials not found in ~/.aws/credentials, fallback to speech-to-text.")
+    
     # Log the conversation
     logger.info("User: {}", initial_prompt)
     logger.info("AI: {}", text)
